fix: reset push to main
