+++ push/index.ts
@@ -1,4 +1,4 @@
-import { $, CryptoHasher, file, sleep, write } from "bun";
+import { $, CryptoHasher, file, write } from "bun";
 import { extract } from "tar";
 
 import stream from "node:stream";
@@ -37,160 +37,196 @@ if (installBun.exitCode !== 0) {
   process.exit(1);
 }
 
-console.log("Preparing image...");
-const imageMetadataRes = await $`docker images --format "{{ .ID }}" ${image}`.quiet();
-if (imageMetadataRes.exitCode !== 0) {
-  console.error(
-    "Image",
-    image,
-    "doesn't exist. The docker daemon might not be running, or the image doesn't exist. Check your existing images with\n\n\tdocker images",
-  );
-  process.exit(1);
-}
 
-const imageID = imageMetadataRes.text();
-if (imageID === "") {
-  console.error("Image", image, "doesn't exist. Check your existing images with\n\n\tdocker images");
-  process.exit(1);
-}
+console.log("Preparing image...");
+// const imageMetadataRes = await $`/usr/bin/docker images --format "{{ .ID }}" ${image}`;
+// if (imageMetadataRes.exitCode !== 0) {
+//   console.error(
+//     "Image",
+//     image,
+//     "doesn't exist. The docker daemon might not be running, or the image doesn't exist. Check your existing images with\n\n\tdocker images",
+//   );
+//   process.exit(1);
+// }
+
+// const imageID = imageMetadataRes.text();
+// if (imageID === "") {
+//   console.error("Image", image, "doesn't exist. Check your existing images with\n\n\tdocker images");
+//   process.exit(1);
+// }
 
 console.log(`Image ${image} found locally, saving to disk...`);
 
-const tarFile = imageID.trim() + ".tar";
+const tarFile = "output.tar";
 const imagePath = ".output-image";
-if (!(await file(tarFile).exists())) {
-  const output = await $`docker save ${image} --output ${tarFile}`;
+if ((await file(tarFile).exists())) {
+  // const output = await $`/usr/bin/docker save ${image} --output ${tarFile}`;
 
-  if (output.exitCode != 0) {
-    console.error("Error saving image", image, output.text());
-    process.exit(1);
-  }
+  // if (output.exitCode != 0) {
+  //   console.error("Error saving image", image, output.text());
+  //   process.exit(1);
+  // }
 
   console.log(`Image saved as ${tarFile}, extracting...`);
 
   await mkdir(imagePath);
 
-  await extract({
+  const result = await extract({
     file: tarFile,
     cwd: imagePath,
   });
 
   console.log(`Extracted to ${imagePath}`);
+} 
+
+type IndexJSONFile = {
+  manifests: {
+    digest: string,
+  }[]
 }
+
+const indexJSONFile = (await Bun.file(path.join(imagePath, "index.json")).json()) as IndexJSONFile;
+// if (indexJSONFile.manifests.length > 0) {
+//   throw new Error('More than one manifest file found')
+// }
+
+
+console.log(indexJSONFile)
+let manifestFile = indexJSONFile.manifests[0].digest
+manifestFile = manifestFile.replace("sha256:", "")
+console.log(manifestFile, "manifest file found");
+
+const manifestBlobFile = file(`${imagePath}/blobs/sha256/${manifestFile}`)
+await Bun.write(`${imagePath}/manifest.json`, manifestBlobFile);
+
+// const moveManifestFileRes = await $`cp ${imagePath}/blobs/sha256/${manifestFile} ${imagePath}/manifest.json`
+// if (moveManifestFileRes.exitCode !== 0) {
+//   throw new Error('Could not move manifest file to manifest.json')
+// }
 
 type DockerSaveConfigManifest = {
-  Config: string;
-  Layers: string[];
-}[];
+  config: {
+    digest: string;
+    size: number;
+  };
+  layers: {
+    digest: string;
+  }[];
+};
 
 import path from "path";
-const manifests = (await Bun.file(path.join(imagePath, "manifest.json")).json()) as DockerSaveConfigManifest;
+const manifest = (await Bun.file(path.join(imagePath, "manifest.json")).json()) as DockerSaveConfigManifest;
 
-if (manifests.length == 0) {
-  console.error("unexpected manifest of length 0");
-  process.exit(1);
-}
+// if (manifests.length == 0) {
+//   console.error("unexpected manifest of length 0");
+//   process.exit(1);
+// }
 
-if (manifests.length > 1) {
-  console.warn("Manifest resolved to multiple images, picking the first one");
-}
+// if (manifests.length > 1) {
+//   console.warn("Manifest resolved to multiple images, picking the first one");
+// }
+
+console.log(manifest, "manifest")
 
 import plimit from "p-limit";
 const pool = plimit(5);
 import zlib from "node:zlib";
 import { rename, rm } from "node:fs/promises";
 
-const cacheFolder = ".cache";
+const cacheFolder = imagePath;
 
 await mkdir(cacheFolder, { recursive: true });
 
-const [manifest] = manifests;
+// const [manifest] = manifests;
 const tasks = [];
 
+console.log(manifest.layers)
 console.log("Compressing...");
 // Iterate through every layer, read it and compress to a file
-for (const layer of manifest.Layers) {
+for (const layer of manifest.layers) {
   tasks.push(
     pool(async () => {
-      let layerPath = path.join(imagePath, layer);
-      // docker likes to put stuff in two ways:
-      //   1. blobs/sha256/<layer>
-      //   2. <layer>/layer.tar
-      //
-      // This handles both cases.
-      let layerName = layer.endsWith(".tar") ? path.dirname(layer) : path.basename(layer);
 
-      const layerCachePath = path.join(cacheFolder, layerName + "-ptr");
-      {
-        const layerCacheGzip = file(layerCachePath);
-        if (await layerCacheGzip.exists()) {
-          const compressedDigest = await layerCacheGzip.text();
-          return compressedDigest;
-        }
-      }
+      return layer.digest.replace("sha256:", "")
+      // let layerPath = path.join(imagePath, layer);
+      // // docker likes to put stuff in two ways:
+      // //   1. blobs/sha256/<layer>
+      // //   2. <layer>/layer.tar
+      // //
+      // // This handles both cases.
+      // let layerName = layer.endsWith(".tar") ? path.dirname(layer) : path.basename(layer);
 
-      const inprogressPath = path.join(cacheFolder, layerName + "-in-progress");
-      await rm(inprogressPath, { recursive: true });
+      // const layerCachePath = path.join(cacheFolder, layerName + "-ptr");
+      // {
+      //   const layerCacheGzip = file(layerCachePath);
+      //   if (await layerCacheGzip.exists()) {
+      //     const compressedDigest = await layerCacheGzip.text();
+      //     return compressedDigest;
+      //   }
+      // }
 
-      const hasher = new Bun.CryptoHasher("sha256");
-      const cacheWriter = file(inprogressPath).writer();
-      const gzipStream = zlib.createGzip({ level: 9 });
-      const writeStream = new stream.Writable({
-        write(val: Buffer, _, cb) {
-          hasher.update(val, "binary");
-          cacheWriter.write(val);
-          cb();
-        },
-      });
+      // const inprogressPath = path.join(cacheFolder, layerName + "-in-progress");
+      // await rm(inprogressPath, { recursive: true });
 
-      gzipStream.pipe(writeStream);
-      await file(layerPath)
-        .stream()
-        .pipeTo(
-          new WritableStream({
-            write(value: Buffer) {
-              return new Promise(async (res) => {
-                const needsWriteBackoff = gzipStream.write(value);
-                // We need to back-off with the writes
-                if (!needsWriteBackoff) {
-                  const onDrain = () => {
-                    // Remove event listener when it finishes
-                    gzipStream.off("drain", onDrain);
-                    res();
-                  };
+      // const hasher = new Bun.CryptoHasher("sha256");
+      // const cacheWriter = file(inprogressPath).writer();
+      // const gzipStream = zlib.createGzip({ level: 9 });
+      // const writeStream = new stream.Writable({
+      //   write(val: Buffer, _, cb) {
+      //     hasher.update(val, "binary");
+      //     cacheWriter.write(val);
+      //     cb();
+      //   },
+      // });
 
-                  gzipStream.on("drain", onDrain);
-                  return;
-                }
+      // gzipStream.pipe(writeStream);
+      // await file(layerPath)
+      //   .stream()
+      //   .pipeTo(
+      //     new WritableStream({
+      //       write(value: Buffer) {
+      //         return new Promise(async (res) => {
+      //           const needsWriteBackoff = gzipStream.write(value);
+      //           // We need to back-off with the writes
+      //           if (!needsWriteBackoff) {
+      //             const onDrain = () => {
+      //               // Remove event listener when it finishes
+      //               gzipStream.off("drain", onDrain);
+      //               res();
+      //             };
 
-                res();
-              });
-            },
-            close() {
-              return new Promise(async (res) => {
-                // Flush before end
-                await new Promise((resFlush) => gzipStream.flush(() => resFlush(true)));
-                // End the stream
-                gzipStream.end(res);
-              });
-            },
-          }),
-        );
+      //             gzipStream.on("drain", onDrain);
+      //             return;
+      //           }
 
-      // Wait until the gzipStream has finished all the piping
-      await new Promise((res) => gzipStream.on("end", () => res(true)));
+      //           res();
+      //         });
+      //       },
+      //       close() {
+      //         return new Promise(async (res) => {
+      //           // Flush before end
+      //           await new Promise((resFlush) => gzipStream.flush(() => resFlush(true)));
+      //           // End the stream
+      //           gzipStream.end(res);
+      //         });
+      //       },
+      //     }),
+      //   );
 
-      const digest = hasher.digest("hex");
-      await rename(inprogressPath, path.join(cacheFolder, digest));
-      await write(layerCachePath, digest);
-      return digest;
+      // // Wait until the gzipStream has finished all the piping
+      // await new Promise((res) => gzipStream.on("end", () => res(true)));
+
+      // const digest = hasher.digest("hex");
+      // await rename(inprogressPath, path.join(cacheFolder, digest));
+      // await write(layerCachePath, digest);
+      // return digest;
     }),
   );
 }
 
-const configManifest = path.join(imagePath, manifest.Config);
-const config = await file(configManifest).text();
-const configDigest = new CryptoHasher("sha256").update(config).digest("hex");
+// const configManifest = manifest.Config
+const config = manifest.config
+const configDigest = config.digest.replace("sha256:", "")
 
 const compressedDigests = await Promise.all(tasks);
 
@@ -250,6 +286,7 @@ async function pushLayer(layerDigest: string, readableStream: ReadableStream, to
     throw new Error(`oci-chunk-max-length header is malformed (not a number)`);
   }
 
+  const reader = readableStream.getReader();
   const uploadId = createUploadResponse.headers.get("docker-upload-uuid");
   if (uploadId === null) {
     throw new Error("Docker-Upload-UUID not defined in headers");
@@ -269,13 +306,9 @@ async function pushLayer(layerDigest: string, readableStream: ReadableStream, to
   let written = 0;
   let previousReadable: ReadableLimiter | undefined;
   let totalLayerSizeLeft = totalLayerSize;
-  const reader = readableStream.getReader();
-  let fail = "true";
-  let failures = 0;
   while (totalLayerSizeLeft > 0) {
     const range = `0-${Math.min(end, totalLayerSize) - 1}`;
     const current = new ReadableLimiter(reader as ReadableStreamDefaultReader, maxToWrite, previousReadable);
-    await current.init();
     const patchChunkUploadURL = parseLocation(location);
     // we have to do fetchNode because Bun doesn't allow setting custom Content-Length.
     // https://github.com/oven-sh/bun/issues/10507
@@ -286,19 +319,14 @@ async function pushLayer(layerDigest: string, readableStream: ReadableStream, to
         "range": range,
         "authorization": cred,
         "content-length": `${Math.min(totalLayerSizeLeft, maxToWrite)}`,
-        "x-fail": fail,
       }),
     });
     if (!patchChunkResult.ok) {
-      previousReadable = current;
-      console.error(`${layerDigest}: Pushing ${range} failed with ${patchChunkResult.status}, retrying`);
-      await sleep(500);
-      if (failures++ >= 2) fail = "false";
-      continue;
+      throw new Error(
+        `uploading chunk ${patchChunkUploadURL} returned ${patchChunkResult.status}: ${await patchChunkResult.text()}`,
+      );
     }
 
-    fail = "true";
-    current.ok();
     const rangeResponse = patchChunkResult.headers.get("range");
     if (rangeResponse !== range) {
       throw new Error(`unexpected Range header ${rangeResponse}, expected ${range}`);
@@ -315,24 +343,18 @@ async function pushLayer(layerDigest: string, readableStream: ReadableStream, to
   const range = `0-${written - 1}`;
   const uploadURL = new URL(parseLocation(location));
   uploadURL.searchParams.append("digest", layerDigest);
-  for (let tries = 0; tries < 3; tries++) {
-    const response = await fetch(uploadURL.toString(), {
-      method: "PUT",
-      headers: new Headers({
-        Range: range,
-        Authorization: cred,
-      }),
-    });
-    if (!response.ok) {
-      console.error(`${layerDigest}: Finishing ${range} failed with ${response.status}, retrying`);
-      continue;
-    }
-
-    console.log("Pushed", layerDigest);
-    return;
+  const response = await fetch(uploadURL.toString(), {
+    method: "PUT",
+    headers: new Headers({
+      Range: range,
+      Authorization: cred,
+    }),
+  });
+  if (!response.ok) {
+    throw new Error(`${uploadURL.toString()} failed with ${response.status}: ${await response.text()}`);
   }
 
-  throw new Error(`Could not push after multiple tries`);
+  console.log("Pushed", layerDigest);
 }
 
 const layersManifest = [] as {
@@ -342,7 +364,7 @@ const layersManifest = [] as {
 }[];
 
 for (const compressedDigest of compressedDigests) {
-  let layer = file(path.join(cacheFolder, compressedDigest));
+  let layer = file(`${imagePath}/blobs/sha256/${compressedDigest}`);
   layersManifest.push({
     mediaType: "application/vnd.oci.image.layer.v1.tar+gzip",
     size: layer.size,
@@ -350,7 +372,7 @@ for (const compressedDigest of compressedDigests) {
   } as const);
   pushTasks.push(
     pool(async () => {
-      const maxRetries = +(process.env["MAX_RETRIES"] ?? 3);
+      const maxRetries = +(process.env["MAX_RETRIES"] ?? 8);
       if (isNaN(maxRetries)) throw new Error("MAX_RETRIES is not a number");
 
       for (let i = 0; i < maxRetries; i++) {
@@ -361,7 +383,7 @@ for (const compressedDigest of compressedDigests) {
           return;
         } catch (err) {
           console.error(digest, "failed to upload", maxRetries - i - 1, "left...", err);
-          layer = file(path.join(cacheFolder, compressedDigest));
+          layer = file(`${imagePath}/blobs/sha256/${compressedDigest}`);
         }
       }
     }),
@@ -370,18 +392,39 @@ for (const compressedDigest of compressedDigests) {
 
 pushTasks.push(
   pool(async () => {
-    await pushLayer(
-      `sha256:${configDigest}`,
-      new ReadableStream({
-        pull(controller) {
-          controller.enqueue(config);
-          controller.close();
-        },
-      }),
-      config.length,
-    );
+    const maxRetries = +(process.env["MAX_RETRIES"] ?? 8);
+    if (isNaN(maxRetries)) throw new Error("MAX_RETRIES is not a number");
+
+    for (let i = 0; i < maxRetries; i++) {
+      let configLayer = file(`${imagePath}/blobs/sha256/${configDigest}`);
+      const stream = configLayer.stream();
+      const digest = `sha256:${configDigest}`;
+      try {
+        await pushLayer(digest, stream, configLayer.size);
+        return;
+      } catch (err) {
+        console.error(digest, "failed to upload", maxRetries - i - 1, "left...", err);
+        configLayer = file(`${imagePath}/blobs/sha256/${configDigest}`);
+      }
+    }
   }),
 );
+
+
+// pushTasks.push(
+//   pool(async () => {
+//     await pushLayer(
+//       `sha256:${configDigest}`,
+//       new ReadableStream({
+//         pull(controller) {
+//           controller.enqueue(config);
+//           controller.close();
+//         },
+//       }),
+//       config.size,
+//     );
+//   }),
+// );
 
 const promises = await Promise.allSettled(pushTasks);
 for (const promise of promises) {
@@ -395,7 +438,7 @@ const manifestObject = {
   mediaType: "application/vnd.oci.image.manifest.v1+json",
   config: {
     mediaType: "application/vnd.oci.image.config.v1+json",
-    size: config.length,
+    size: config.size,
     digest: `sha256:${configDigest}`,
   },
   layers: layersManifest,
